Poopulated index, started working on the UI

+++ Prj/ECivis/ECivisObj/RoboticEntDetails.cs
@@ -15,7 +15,7 @@ namespace ECivisObj
         public string Category { get; set; }
         public string Description { get; set; }
         public long MemberCount { get; set; }
-        public int? Rating { get; set; }
+        public double? Rating { get; set; }
         public string ContactDetails { get; set; }
         public double? WomenPercentage { get; set; }
         public int? Sentiment { get; set; }
